# docs/spec/AEGIS_API_SPEC.yaml
openapi: 3.1.0
info:
  title: AEGIS Agent Identity & Verification API
  version: 1.0.0
  description: |
    The AEGIS API provides verified identity, scoped authorization, and behavioral
    attestation for AI agents. Neutral, platform-agnostic, ACP-compatible.
  contact:
    email: dev@aegislabs.io
  license:
    name: Proprietary

servers:
  - url: https://api.aegislabs.io/v1
    description: Production
  - url: https://sandbox.aegislabs.io/v1
    description: Sandbox

tags:
  - name: Identity
    description: Agent registration and identity management
  - name: Policies
    description: Scoped permission policies
  - name: Verification
    description: Real-time agent verification (relying party endpoints)
  - name: Audit
    description: Audit log access
  - name: Reporting
    description: Behavioral signal reporting by relying parties

paths:
  # ─────────────────────────────────────────────────
  # IDENTITY
  # ─────────────────────────────────────────────────

  /agents/register:
    post:
      tags: [Identity]
      summary: Register a new agent identity
      description: |
        Creates a new agent identity record. The developer provides their Ed25519
        public key; AEGIS never holds the private key. Returns an agentId and
        a verification token for confirming the keypair.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegistrationRequest'
      responses:
        '201':
          description: Agent registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRegistrationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /agents/{agentId}:
    get:
      tags: [Identity]
      summary: Get agent identity details
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentIdentity'
          description: Agent details
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Identity]
      summary: Revoke an agent identity
      description: Permanently revokes the agent. Cannot be undone.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '204':
          description: Agent revoked
        '404':
          $ref: '#/components/responses/NotFound'

  /agents/{agentId}/status:
    get:
      tags: [Identity]
      summary: Get agent status and trust score (public endpoint)
      description: |
        Public endpoint — no API key required. Returns current status and
        trust score for a given agentId. Used by relying parties for quick
        pre-checks before calling /verify.
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentStatus'
          description: Agent status

  # ─────────────────────────────────────────────────
  # POLICIES
  # ─────────────────────────────────────────────────

  /agents/{agentId}/policies:
    post:
      tags: [Policies]
      summary: Create a scoped permission policy for an agent
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyCreateRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyCreateResponse'
          description: Policy created

    get:
      tags: [Policies]
      summary: List active policies for an agent
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentPolicy'
          description: Active policies

  /agents/{agentId}/policies/{policyId}:
    delete:
      tags: [Policies]
      summary: Revoke a policy (instant)
      description: |
        Immediately invalidates the policy. Any subsequent /verify call
        referencing this policy will return valid: false.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - $ref: '#/components/parameters/PolicyId'
      responses:
        '204':
          description: Policy revoked

  # ─────────────────────────────────────────────────
  # VERIFICATION
  # ─────────────────────────────────────────────────

  /verify:
    post:
      tags: [Verification]
      summary: Verify an inbound agent token
      description: |
        Primary relying-party endpoint. Verifies the cryptographic signature,
        checks policy scope, validates spend limits, and returns the current
        BATE trust score. Target SLA: <80ms p99.
      security:
        - ApiKeyAuth: []
        - PublicVerifyKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
          description: Verification result (check `valid` field — 200 even when invalid)
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ─────────────────────────────────────────────────
  # AUDIT
  # ─────────────────────────────────────────────────

  /agents/{agentId}/audit:
    get:
      tags: [Audit]
      summary: Retrieve audit log for an agent
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 1000
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogResponse'
          description: Audit events

  # ─────────────────────────────────────────────────
  # REPORTING (relying party → BATE signal ingestion)
  # ─────────────────────────────────────────────────

  /agents/{agentId}/report:
    post:
      tags: [Reporting]
      summary: Report a behavioral signal about an agent
      description: |
        Relying parties use this to report fraud, anomalies, or policy
        violations. Signals feed directly into BATE and affect trust score.
        Verified relying parties are weighted more heavily.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '202':
          description: Report accepted (async processing)

# ─────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-AEGIS-API-Key
    PublicVerifyKey:
      type: apiKey
      in: header
      name: X-AEGIS-Verify-Key
      description: Separate key for relying parties (read-only verify access)

  parameters:
    AgentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
        example: "agt_01HZ9YZXM4QT3B7P8WKJD6R5V"
    PolicyId:
      name: policyId
      in: path
      required: true
      schema:
        type: string
        example: "pol_01HZ9YZXM4QT3B7P8WKJD6R5V"

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:

    AgentRegistrationRequest:
      type: object
      required: [publicKey, runtime, principalId]
      properties:
        publicKey:
          type: string
          description: Ed25519 public key (base64url encoded)
          example: "MCowBQYDK2VwAyEA..."
        runtime:
          type: string
          enum: [openai, anthropic, google, custom]
        model:
          type: string
          example: "gpt-4o"
        principalId:
          type: string
          description: Your AEGIS principal (human/org) ID
        label:
          type: string
          description: Human-readable label for this agent
          example: "Shopping agent for erwin@klytics.io"

    AgentRegistrationResponse:
      type: object
      properties:
        agentId:
          type: string
          example: "agt_01HZ9YZXM4QT3B7P8WKJD6R5V"
        verificationToken:
          type: string
          description: Sign this with your private key to confirm keypair ownership
        trustScore:
          type: integer
          example: 500
          description: Starting score (500 = neutral)
        registeredAt:
          type: string
          format: date-time

    AgentIdentity:
      type: object
      properties:
        agentId:
          type: string
        publicKey:
          type: string
        principalId:
          type: string
        runtime:
          type: string
        model:
          type: string
        label:
          type: string
        status:
          type: string
          enum: [active, suspended, revoked]
        trustScore:
          type: integer
          minimum: 0
          maximum: 1000
        trustBand:
          type: string
          enum: [PLATINUM, VERIFIED, WATCH, FLAGGED]
        registeredAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time

    AgentStatus:
      type: object
      properties:
        agentId:
          type: string
        status:
          type: string
          enum: [active, suspended, revoked]
        trustScore:
          type: integer
        trustBand:
          type: string
          enum: [PLATINUM, VERIFIED, WATCH, FLAGGED]
        lastSeenAt:
          type: string
          format: date-time

    PolicyCreateRequest:
      type: object
      required: [scopes, expiresAt]
      properties:
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/PolicyScope'
        expiresAt:
          type: string
          format: date-time
        label:
          type: string
          example: "Book flights under $500"

    PolicyCreateResponse:
      type: object
      properties:
        policyId:
          type: string
        signedToken:
          type: string
          description: JWT signed by AEGIS — attach to agent outbound requests
        expiresAt:
          type: string
          format: date-time

    AgentPolicy:
      type: object
      properties:
        policyId:
          type: string
        agentId:
          type: string
        scopes:
          type: array
          items:
            $ref: '#/components/schemas/PolicyScope'
        status:
          type: string
          enum: [active, expired, revoked]
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time

    PolicyScope:
      type: object
      required: [category]
      properties:
        category:
          type: string
          enum: [commerce, data-read, data-write, communication, scheduling]
        spendLimit:
          type: object
          properties:
            currency:
              type: string
              enum: [USD, EUR, GBP]
            maxPerTransaction:
              type: number
              example: 500
            maxPerDay:
              type: number
              example: 1000
            maxPerMonth:
              type: number
              example: 5000
        merchantCategories:
          type: array
          items:
            type: string
          description: MCC codes (e.g. ["3000-3299"] for airlines)
          example: ["3000-3299", "5411"]
        allowedDomains:
          type: array
          items:
            type: string
          example: ["delta.com", "united.com"]
        dataScopes:
          type: array
          items:
            type: string
          example: ["read:email", "write:calendar"]
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time

    VerifyRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Signed AEGIS token from the agent (JWT format)
        action:
          type: string
          example: "commerce.purchase"
        amount:
          type: number
          example: 450.00
        currency:
          type: string
          example: "USD"
        merchantId:
          type: string
          example: "delta-airlines"
        merchantDomain:
          type: string
          example: "delta.com"
        context:
          type: object
          additionalProperties: true
          description: Additional context for BATE scoring

    VerifyResponse:
      type: object
      properties:
        valid:
          type: boolean
        agentId:
          type: string
        principalId:
          type: string
        trustScore:
          type: integer
          minimum: 0
          maximum: 1000
        trustBand:
          type: string
          enum: [PLATINUM, VERIFIED, WATCH, FLAGGED]
        scopesGranted:
          type: array
          items:
            type: string
          example: ["commerce", "commerce.purchase"]
        spendRemaining:
          type: object
          properties:
            today:
              type: number
            thisMonth:
              type: number
        denialReason:
          type: string
          nullable: true
          description: |
            Only present when valid=false. The enum order below is the
            CANONICAL DENIAL PRECEDENCE (top wins) — see CLAUDE.md
            invariant 6 / ADR-0004 / packages/types DENIAL_REASON_PRECEDENCE.
            Relying-party SDKs depend on this order; do not reorder
            without a major API version bump.
            PLAN_LIMIT_EXCEEDED is a billing pre-gate that fires before the
            9-step algorithm chain — it is always position 0 in this enum.
          enum:
            - PLAN_LIMIT_EXCEEDED
            - AGENT_NOT_FOUND
            - AGENT_REVOKED
            - INVALID_SIGNATURE
            - POLICY_REVOKED
            - POLICY_EXPIRED
            - SCOPE_NOT_GRANTED
            - TRIAL_EXHAUSTED
            - SPEND_LIMIT_EXCEEDED
            - TRUST_SCORE_TOO_LOW
            - ANOMALY_FLAGGED
        verifiedAt:
          type: string
          format: date-time
        ttl:
          type: integer
          description: Seconds this result can be cached by relying party
          example: 30

    AuditEvent:
      type: object
      properties:
        eventId:
          type: string
        agentId:
          type: string
        principalId:
          type: string
        timestamp:
          type: string
          format: date-time
        action:
          type: string
        relyingParty:
          type: string
        decision:
          type: string
          enum: [approved, denied, flagged]
        decisionReason:
          type: string
          nullable: true
        trustScoreAtEvent:
          type: integer
        signature:
          type: string
          description: AEGIS signature over this event (immutability proof)

    AuditLogResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/AuditEvent'
        nextCursor:
          type: string
          nullable: true
        total:
          type: integer

    ReportRequest:
      type: object
      required: [eventType]
      properties:
        eventType:
          type: string
          enum: [fraud_confirmed, anomaly, policy_violation, suspicious_behavior, false_positive]
        severity:
          type: string
          enum: [low, medium, high, critical]
          default: medium
        description:
          type: string
          maxLength: 1000
        transactionId:
          type: string
          description: Your internal transaction ID for correlation
        evidence:
          type: object
          additionalProperties: true
          description: Any additional signal data (IP, timestamps, amounts)

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
        requestId:
          type: string